--- Assets/Prefabs/Environment/Fish/EmperorAngelfish.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c3fcd2297df196348ad8d3ab01c6eb46
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
