--- Assets/Prefabs/PalmTree.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ea5e0575bd1a18e469dcb1c97b8b6f63
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
