+++ Assets/Prefabs/Spear.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 3d2d5a8d514734c4281a9b2947ee204f
-PrefabImporter:
+guid: c123403b6b14c284089c13f681b3d5ff
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
