+++ Assets/Scripts/NPCs/Humans.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 2be16683c31311844a0376849c424e75
-PrefabImporter:
+guid: 3d4ac4daf238b81479f1c100e8b0c014
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
