+++ Assets/Scripts/NPCs/Humans/Items.meta
