+++ Assets/Scripts/NPCs/Humans/Items/Rope.cs.meta
