+++ Assets/Scripts/NPCs/Humans/ThrowSpear.cs.meta
